Linux Bash Scripting For MCQs
In this project, we are simply developing a script for students where they will add their MCQ answers. In this task, we are trying to implement loops and if function whether if any false option is added they might receive an error which will cause the MCQs to be incorrect.