Linux Bash Scripting
In this project, we are trying to build up a grade checker which will determine whether the student is passing or failing and which student has the highest and the lowest grade. We are implementing an IFS(internal field separator) and if functions, to meet the argument that if the student's result is greater or lower than 0-100 than output will return as an error message.