Linux Bash Scripting For Grade Checker
In this project, we are trying to build up a grade checker where we will add the student's grades and find the class average and also which student has the highest and the lowest grade. We are implementing an IFS(internal field separator) and if functions, to meet the argument that if the student's result is greater or lower than 0-100 then the output will return as an error message.